Improving Pinoccio Scout port. Testing new avr tools.
1. avr-g++ (AVR_8_bit_GNU_Toolchain_3.4.3_1072) 4.8.1, needs -O3 to
deliver same as previous.
2. Removed warnings.
3. Some new conflicting symbols.

+++ cores/cosa/Cosa/INET/SNMP.cpp
@@ -359,7 +359,7 @@ SNMP::encode_string(const char* buf)
 bool 
 SNMP::encode_sequence(int32_t count)
 {
-  uint8_t header[] = { SYNTAX_SEQUENCE, count };
+  uint8_t header[] = { SYNTAX_SEQUENCE, (uint8_t) count };
   return (write(header, sizeof(header)) == sizeof(header));
 }
